Support cancellation on single/scalar async ops
Motivation
----------
When using FirstAsync/SingleAsync or similar ops, the cancellation token
should be used rather than ignored.

Modifications
-------------
Rework AsyncStreamValue, AsyncStreamedValueInfo, and
AsyncStreamedSingleValueInfo to pass along the cancellation token when
provided.

Results
-------
Single/scalar async operations are cancellable.
