+++ Src/Couchbase.Linq/Execution/StreamedData/AsyncStreamedSingleValueInfo.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 using System.Reflection;
+using System.Threading;
+using System.Threading.Tasks;
 using Remotion.Linq;
 using Remotion.Linq.Clauses.StreamedData;
 
@@ -11,10 +13,6 @@ namespace Couchbase.Linq.Execution.StreamedData
     /// </summary>
     internal class AsyncStreamedSingleValueInfo : AsyncStreamedValueInfo
     {
-        private static readonly MethodInfo ExecuteMethod =
-            typeof(AsyncStreamedSingleValueInfo).GetMethod(nameof(ExecuteSingleQueryModel),
-                new[] {typeof(QueryModel), typeof(IClusterQueryExecutor)});
-
         public bool ReturnDefaultWhenEmpty { get; }
 
         public AsyncStreamedSingleValueInfo(Type dataType, bool returnDefaultWhenEmpty)
@@ -23,45 +21,23 @@ namespace Couchbase.Linq.Execution.StreamedData
             ReturnDefaultWhenEmpty = returnDefaultWhenEmpty;
         }
 
-        public override IStreamedData ExecuteQueryModel(QueryModel queryModel, IQueryExecutor executor)
-        {
-            if (queryModel == null)
-            {
-                throw new ArgumentNullException(nameof(queryModel));
-            }
-            if (executor == null)
-            {
-                throw new ArgumentNullException(nameof(executor));
-            }
-            if (!(executor is IClusterQueryExecutor asyncExecutor))
-            {
-                throw new ArgumentException($"{nameof(executor)} must implement {typeof(IClusterQueryExecutor)} for asynchronous queries.");
-            }
-
-            var executeMethod = ExecuteMethod.MakeGenericMethod(InternalType);
-
-            // wrap executeMethod into a delegate instead of calling Invoke in order to allow for exceptions that are bubbled up correctly
-            var func = (Func<QueryModel, IClusterQueryExecutor, object>) executeMethod.CreateDelegate (typeof (Func<QueryModel, IClusterQueryExecutor, object>), this);
-            var result = func(queryModel, asyncExecutor);
-
-            return new AsyncStreamedValue(result, this);
-        }
-
         protected override AsyncStreamedValueInfo CloneWithNewDataType(Type dataType) =>
             new AsyncStreamedSingleValueInfo (dataType, ReturnDefaultWhenEmpty);
 
-        public object ExecuteSingleQueryModel<T>(QueryModel queryModel, IClusterQueryExecutor executor)
+        public override Task<T> ExecuteQueryModelAsync<T>(QueryModel queryModel, IClusterQueryExecutor executor,
+            CancellationToken cancellationToken = default)
         {
             if (queryModel == null)
             {
                 throw new ArgumentNullException(nameof(queryModel));
             }
+
             if (executor == null)
             {
                 throw new ArgumentNullException(nameof(executor));
             }
 
-            return executor.ExecuteSingleAsync<T>(queryModel, ReturnDefaultWhenEmpty);
+            return executor.ExecuteSingleAsync<T>(queryModel, ReturnDefaultWhenEmpty, cancellationToken);
         }
 
         // ReSharper disable PossibleNullReferenceException
